requirements.txt
pydng==1.1.2
pydantic==1.7.3
requests~=2.25.1
fastapi~=0.63.0
PyGithub==1.54.1
tinydb~=4.3.0
PyNaCl==1.4.0
uvicorn==0.13.4
